Move graph helpers to deploy/api
