+++ pkg/deploy/graph/helpers.go
@@ -10,7 +10,6 @@ import (
 	deployapi "github.com/openshift/origin/pkg/deploy/api"
 	deploygraph "github.com/openshift/origin/pkg/deploy/graph/nodes"
 	deployutil "github.com/openshift/origin/pkg/deploy/util"
-	imageapi "github.com/openshift/origin/pkg/image/api"
 )
 
 // RelevantDeployments returns the active deployment and a list of inactive deployments (in order from newest to oldest)
@@ -47,59 +46,4 @@ func (m RecentDeploymentReferences) Len() int      { return len(m) }
 func (m RecentDeploymentReferences) Swap(i, j int) { m[i], m[j] = m[j], m[i] }
 func (m RecentDeploymentReferences) Less(i, j int) bool {
 	return deployutil.DeploymentVersionFor(m[i].ReplicationController) > deployutil.DeploymentVersionFor(m[j].ReplicationController)
-}
-
-// TODO: move to deploy/api/helpers.go
-type TemplateImage struct {
-	Image string
-
-	Ref *imageapi.DockerImageReference
-
-	From *kapi.ObjectReference
-}
-
-func EachTemplateImage(pod *kapi.PodSpec, triggerFn TriggeredByFunc, fn func(TemplateImage, error)) {
-	for _, container := range pod.Containers {
-		var ref imageapi.DockerImageReference
-		if trigger, ok := triggerFn(&container); ok {
-			trigger.Image = container.Image
-			fn(trigger, nil)
-			continue
-		}
-		ref, err := imageapi.ParseDockerImageReference(container.Image)
-		if err != nil {
-			fn(TemplateImage{Image: container.Image}, err)
-			continue
-		}
-		fn(TemplateImage{Image: container.Image, Ref: &ref}, nil)
-	}
-}
-
-type TriggeredByFunc func(container *kapi.Container) (TemplateImage, bool)
-
-func DeploymentConfigHasTrigger(config *deployapi.DeploymentConfig) TriggeredByFunc {
-	return func(container *kapi.Container) (TemplateImage, bool) {
-		for _, trigger := range config.Spec.Triggers {
-			params := trigger.ImageChangeParams
-			if params == nil {
-				continue
-			}
-			for _, name := range params.ContainerNames {
-				if container.Name == name {
-					if len(params.From.Name) == 0 {
-						continue
-					}
-					from := params.From
-					if len(from.Namespace) == 0 {
-						from.Namespace = config.Namespace
-					}
-					return TemplateImage{
-						Image: container.Image,
-						From:  &from,
-					}, true
-				}
-			}
-		}
-		return TemplateImage{}, false
-	}
 }
